pyproject.toml
[project]
name = "psstoreprice"
version = "0.1.0"
description = "Aplikacja do monitorowania cen w PS Store"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-auth==2.49.1",
    "gspread==6.2.1",
    "python-dotenv==1.2.2",
]

[tool.uv]
package = false
